requirements.txt
# Απαιτήσεις για το Cinema Django Backend
# Requirements for Cinema Django Backend

Django>=5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.0
tmdbv3api>=1.9.0
dependency-injector>=4.41.0

# Testing (integration tests)
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=5.0.0

# MySQL Driver — install the one for your platform:
#   Windows:  pip install PyMySQL
#   Linux:    pip install mysqlclient (requires libmysqlclient-dev)
mysqlclient>=2.2.0; platform_system != "Windows"
PyMySQL>=1.1.0
